requirements.txt
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
dataclasses==0.8
Flask==2.0.1
future==0.18.2
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gspread==3.7.0
httplib2==0.19.1
idna==3.2
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
line-bot-sdk==1.19.0
MarkupSafe==2.0.1
oauth2client==4.1.3
oauthlib==3.1.1
packaging==21.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.5.0
